refactoring ensured base64 audio is decoded into memory before playing
